+++ lib/bpm/package.rb
@@ -236,8 +236,7 @@ module BPM
       begin
         json = JSON.parse(File.read(@json_path))
       rescue JSON::ParserError, Errno::EACCES, Errno::ENOENT => ex
-        add_error "There was a problem parsing #{File.basename(@json_path)}: #{ex.message}"
-        return false
+        raise BPM::InvalidPackageError.new self, ex.message
       end
 
       @read_keys = json.keys.dup # to retain order on save